software/pc_monitor/requirements.txt
PySide6>=6.5.0
pyqtgraph>=0.13.0
pyserial>=3.5
numpy>=1.24.0
